Make sure db_insert_id() always returns an int
db_result() returns a string in some cases. Typecasting the return
value to int ensures we comply with the PHPDoc.

Fixes #22904

+++ core/database_api.php
@@ -567,7 +567,7 @@ function db_insert_id( $p_table = null, $p_field = 'id' ) {
 		}
 		if( isset( $t_query ) ) {
 			$t_result = db_query( $t_query );
-			return db_result( $t_result );
+			return (int)db_result( $t_result );
 		}
 	}
 	return $g_db->Insert_ID();
